+++ src/logging.rs
@@ -7,7 +7,7 @@ use crate::prelude::*;
 pub fn init(opts: &Opts) -> Result {
     let mut config_builder = ConfigBuilder::new();
     config_builder
-        .set_thread_level(LevelFilter::Error)
+        .set_thread_level(LevelFilter::Off)
         .set_target_level(LevelFilter::Error)
         .set_location_level(LevelFilter::Debug)
         .set_thread_mode(ThreadLogMode::Names)