+++ src/sentry.rs
@@ -1,0 +1,7 @@
+use sentry::integrations::log::LogIntegration;
+use sentry::{ClientInitGuard, ClientOptions};
+
+/// Initialize Sentry integration.
+pub fn init(dsn: impl AsRef<str>) -> ClientInitGuard {
+    sentry::init(Into::<ClientOptions>::into(dsn.as_ref()).add_integration(LogIntegration::default()))
+}