+++ src/services/helpers/handle_result.rs
@@ -1,8 +1,7 @@
 use crate::prelude::*;
 
-pub async fn handle_service_result(service_id: &str, sleep_interval: Duration, result: Result) {
-    if let Err(error) = result {
-        error!("[{}] {}", service_id, error);
-    }
-    task::sleep(sleep_interval).await;
+/// Log the result and sleep for the specified duration.
+pub async fn handle_service_result(service_id: &str, sleep_duration: Duration, result: Result) {
+    let _ = result.log(|| format!("[{}]", service_id));
+    task::sleep(sleep_duration).await;
 }