+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # `master`
 
+- 💥 Move some command-line options to the settings
 - 🐛 Fix `select_sensor_reading_count`
 - 🔇 Remove thread IDs from the logs